"changed 3-5"

+++ 2015/s2/oop/practical-01/program3-5.cpp
@@ -2,7 +2,7 @@
 #include <math.h>
 
 
-double sumeven(int array[],int n){
+double sumeven(double array[],int n){
 	int sum=0;
 	int i=0;
 	for(i=0;i<n;i++){
@@ -15,7 +15,7 @@ return sum;
 
 #ifndef WEBSUBMIT
 int main(){
-	int array[5]={1,2,3,2,1};
+	double array[5]={1,2,3,2,1};
 
 }
 #endif //WEBSUBMIT
